+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Documentation of release versions of `nacc-attribute-deriver`
 
 * Update to include non-prescription V1 drugs (e.g. over-the-counter and vitamins - was previously only keeping track of prescription drugs)
 * Update to handle new B1a scope - like MEDS, comes from a separate file and has to be mapped to UDS
+* Fixes issue of values entered in an A1 FVP visit overwriting IVP, when they should not be provided in the first place. Updates to ignore FVP values
 * If a header formdate cannot be determined, return blank for missingness instead of throwing error (handles optional forms)
 * A few other minor fixes from regression testing